FizzBuzz/src/test/java/FizzBuzzTest.java
import org.junit.Test;
import za.co.wethinkcode.FizzBuzz;

import static org.junit.Assert.*;

public class FizzBuzzTest {

    // Test FizzBuzz
    @Test
    public void testFizzBuzz() {
        assertEquals("FizzBuzz", FizzBuzz.fizzBuzz(15));
    }

    // Test Fizz
    @Test
    public void testFizz() {
        assertEquals("Fizz", FizzBuzz.fizzBuzz(3));
        assertEquals("Fizz", FizzBuzz.fizzBuzz(6));
    }

    // Test Buzz
    @Test
    public void testBuzz() {
        assertEquals("Buzz", FizzBuzz.fizzBuzz(5));
        assertEquals("Buzz", FizzBuzz.fizzBuzz(10));
    }

    @Test
    public void testNotFizzOrBuzz(){
        assertEquals("1", FizzBuzz.fizzBuzz(1));
        assertEquals("2", FizzBuzz.fizzBuzz(2));

    }
}
